requirements.txt
transformers==4.40.0
torch==2.2.0
streamlit==1.35.0
pandas==2.2.0
plotly==5.20.0
requests==2.31.0
python-dotenv==1.0.0
yfinance==0.2.40
pymupdf==1.24.0